+++ docs/static/server.swagger.json
@@ -39,7 +39,8 @@
         "NACK",
         "NO_DATA_TO_SIGN",
         "DATA_TO_SIGN",
-        "SIGNER_NOTIFY"
+        "SIGNER_NOTIFY",
+        "RAW_DATA"
       ],
       "default": "KEYGEN",
       "title": "- ACK: it could be a separate MsgType but it would clone logic of requests"
